+++ CHANGELOG.md
@@ -2,4 +2,10 @@
 
 ## unreleased
 
+Added
+-----
+
+* Added Kafka server
+
+
 * Bootstrap repo with ska-cookiecutter-pypackage